validator fixes
