+++ src/scripts/validatorTitle.js
@@ -5,8 +5,5 @@ export default function(validator) {
   } else {
     title = "Anonymous"
   }
-  let shortOwner = validator.owner.split(1)[1]
-  shortOwner = shortOwner.slice(0, 8)
-  title += ` - (${shortOwner})`
   return title
 }
